fix: move passwordSettingsHeading.tsx to own folder in components folder
